test trait delete and correction method getTwig
