+++ core/DefaultAbstractController.php
@@ -3,13 +3,9 @@
 namespace Core;
 
 use Exception;
-use Twig\Environment;
-use Twig\Extension\DebugExtension;
-use Twig\Loader\FilesystemLoader;
 
 abstract class DefaultAbstractController implements DefaultControllerInterface
 {
-    protected static $twig;
     protected        $request;
 
     public function __construct()
@@ -26,7 +22,7 @@ abstract class DefaultAbstractController implements DefaultControllerInterface
     {
         $defaultPath = dirname(__DIR__) . DIRECTORY_SEPARATOR . 'template' . DIRECTORY_SEPARATOR;
         $viewFolder  = $viewFolder ?? $this->getFolderView();
-        $view        = $this->getTwig()->render($viewFolder . $viewName, $params);
+        $view        = (new TwigProvider())->getTwig()->render($viewFolder . $viewName, $params);
 
         //check if the view exist or return of exception
         if (false === file_exists($defaultPath . $viewFolder . $viewName)) {
@@ -40,25 +36,5 @@ abstract class DefaultAbstractController implements DefaultControllerInterface
     {
         // Define the view directory
         return 'front/';
-    }
-
-    public static function getTwig()
-    {
-        // instance of Twig
-        if (null === static::$twig) {
-            $loader       = new FilesystemLoader('template/');
-            static::$twig = new Environment($loader,
-                // To the prod define the path of directory Cache, else to dev keep false
-                                            [
-                                                'cache' => false,
-                                                'debug' => DEBUG_TWIG //ou true
-                                            ]
-            );
-        }
-        if (DEBUG_TWIG) {
-            static::$twig->addExtension(new DebugExtension());
-        }
-
-        return static::$twig;
     }
 }
